Finally fixed messy merge and corrected integer numpy deprecation warnings

+++ esabin/esagrid.py
@@ -96,7 +96,7 @@ class Esagrid(object):
         self.n_lat_bins = int(180./self.delta_lat-1.)
         
         #Create the edges for each bin
-        self.lat_bin_edges = np.linspace(-90.,90.,num=self.n_lat_bins+1.)
+        self.lat_bin_edges = np.linspace(-90.,90.,num=self.n_lat_bins+1)
         self.lon_bin_edges = []
         self.flat_bins = [] #A flat list of bins from south to north, -180 to 180 lon
                             #format is flat_bins[i] = [ith_bin_lat_start,ith_bin_lat_end,ith_bin_lon_start,ith_bin_lon_end]
@@ -310,7 +310,7 @@ class Esagrid(object):
         th1 = (90.-lat_start)*np.pi/180.
         th2 = (90.-lat_end)*np.pi/180.
         N12 = (np.cos(th1)-np.cos(th2))/(1-np.cos(th2-th1))*n_cap_bins
-        N12 = np.abs(np.round(N12))
+        N12 = int(np.abs(np.round(N12)))
         #+1 because we want N12 bins so we need N12+1 edges
         bins = np.linspace(-1*np.pi,np.pi,num=N12+1)*self.azi_fac
